+++ packages/event-bus/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @solid-primitives/event-bus
 
+## 0.1.6
+
+### Patch Changes
+
+- c2866ea6: Update utils package
+- Updated dependencies [c2866ea6]
+  - @solid-primitives/utils@5.0.0
+  - @solid-primitives/immutable@0.1.6
+
 ## 0.1.5
 
 ### Patch Changes