+++ packages/graphql/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @solid-primitives/graphql
 
+## 1.5.2
+
+### Patch Changes
+
+- Updated dependencies [c2866ea6]
+  - @solid-primitives/utils@5.0.0
+
 ## 1.5.1
 
 ### Patch Changes