+++ packages/scroll/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @solid-primitives/scroll
 
+## 2.0.8
+
+### Patch Changes
+
+- Updated dependencies [c2866ea6]
+- Updated dependencies [c2866ea6]
+  - @solid-primitives/event-listener@2.2.6
+  - @solid-primitives/rootless@1.2.3
+  - @solid-primitives/utils@5.0.0
+
 ## 2.0.7
 
 ### Patch Changes